+++ src/test/java/net/kaczmarzyk/spring/data/jpa/domain/EqualIgnoreCaseIntegrationTest.java
@@ -21,11 +21,14 @@ import org.junit.Test;
 import java.util.List;
 
 import static org.assertj.core.api.Assertions.assertThat;
+import static org.junit.Assert.assertThrows;
 
 /**
  * @author Ricardo Pardinho
  */
 public class EqualIgnoreCaseIntegrationTest extends EqualIntegrationTest {
+
+    private static final String INVALID_PARAMETER_ARRAY_SIZE_EXCEPTION_MESSAGE = "Invalid size of 'httpParamValues' array, Expected 1 but was ";
 
     @Test
     public void filtersByStringCaseInsensitive() {
@@ -59,14 +62,31 @@ public class EqualIgnoreCaseIntegrationTest extends EqualIntegrationTest {
         assertThat(moeFound).hasSize(2).containsOnly(homerSimpson, moeSzyslak);
     }
 
-    @Test(expected = IllegalArgumentException.class)
-    public void rejectsMissingArguments() {
-        new EqualIgnoreCase<>(queryCtx, "path", new String[] {}, defaultConverter);
+    @Test
+    public void rejectsNullArgumentArray() {
+        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
+                () -> new EqualIgnoreCase<>(queryCtx, "path", null, defaultConverter));
+
+        assertThat(exception.getMessage())
+                .isEqualTo(INVALID_PARAMETER_ARRAY_SIZE_EXCEPTION_MESSAGE + "null");
     }
 
-    @Test(expected = IllegalArgumentException.class)
+    @Test
+    public void rejectsMissingArguments() {
+        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
+                () -> new EqualIgnoreCase<>(queryCtx, "path", new String[] {}, defaultConverter));
+
+        assertThat(exception.getMessage())
+                .isEqualTo(INVALID_PARAMETER_ARRAY_SIZE_EXCEPTION_MESSAGE + "[]");
+    }
+
+    @Test
     public void rejectsTooManyArguments() {
-        new EqualIgnoreCase<>(queryCtx, "path", new String[] {"2014-03-10", "2014-03-11"}, defaultConverter);
+        IllegalArgumentException exception = assertThrows(IllegalArgumentException.class,
+                () -> new EqualIgnoreCase<>(queryCtx, "path", new String[] {"2014-03-10", "2014-03-11"}, defaultConverter));
+
+        assertThat(exception.getMessage())
+                .isEqualTo(INVALID_PARAMETER_ARRAY_SIZE_EXCEPTION_MESSAGE + "[2014-03-10, 2014-03-11]");
     }
     
 }